+++ changelog.md
@@ -1,2 +1,9 @@
+# 1.0.2
+- Added level info popups support
+- Fixed daily label position
+
+# 1.0.1
+- Added tags
+
 # 1.0.0
 - Initial release